Replace KubeObjectDetailRegistry with injectable version
