+++ package.json
@@ -62,7 +62,8 @@
     },
     "moduleNameMapper": {
       "\\.(css|scss)$": "<rootDir>/__mocks__/styleMock.ts",
-      "\\.(svg)$": "<rootDir>/__mocks__/imageMock.ts"
+      "\\.(svg)$": "<rootDir>/__mocks__/imageMock.ts",
+      "^monaco-editor": "<rootDir>/node_modules/monaco-editor"
     },
     "modulePathIgnorePatterns": [
       "<rootDir>/dist",
@@ -340,6 +341,7 @@
     "esbuild": "^0.13.15",
     "esbuild-loader": "^2.16.0",
     "eslint": "^7.32.0",
+    "eslint-import-resolver-typescript": "^2.5.0",
     "eslint-plugin-header": "^3.1.1",
     "eslint-plugin-import": "^2.25.3",
     "eslint-plugin-react": "^7.27.1",
