+++ src/extensions/registries/index.ts
@@ -25,7 +25,6 @@ export * from "./page-registry";
 export * from "./page-menu-registry";
 export * from "./app-preference-registry";
 export * from "./status-bar-registry";
-export * from "./kube-object-detail-registry";
 export * from "./kube-object-menu-registry";
 export * from "./kube-object-status-registry";
 export * from "./entity-setting-registry";
