+++ src/renderer/bootstrap.tsx
@@ -109,9 +109,6 @@ export async function bootstrap(comp: () => Promise<AppComponent>, di: Dependenc
   logger.info(`${logPrefix} initializing KubeObjectMenuRegistry`);
   initializers.initKubeObjectMenuRegistry();
 
-  logger.info(`${logPrefix} initializing KubeObjectDetailRegistry`);
-  initializers.initKubeObjectDetailRegistry();
-
   logger.info(`${logPrefix} initializing WorkloadsOverviewDetailRegistry`);
   initializers.initWorkloadsOverviewDetailRegistry();
