+++ src/renderer/components/layout/top-bar/top-bar.test.tsx
@@ -90,7 +90,6 @@ describe("<TopBar/>", () => {
 
   beforeEach(() => {
     di = getDiForUnitTesting();
-
     render = renderFor(di);
   });
