+++ src/renderer/initializers/index.ts
@@ -23,7 +23,6 @@ export * from "./catalog-entity-detail-registry";
 export * from "./catalog";
 export * from "./entity-settings-registry";
 export * from "./ipc";
-export * from "./kube-object-detail-registry";
 export * from "./kube-object-menu-registry";
 export * from "./registries";
 export * from "./workloads-overview-detail-registry";
